refactor(core): D2 — env-resolution dedup (P5)
D2 slice shipped. jackin-core::env_model is now the single canonical
home for the env-resolution vocabulary (constants, is_reserved,
extract_interpolation_refs, topological_env_order). The jackin binary's
shim module crates/jackin/src/env_model.rs and its tests file
crates/jackin/src/env_model/tests.rs are deleted; the binary now
imports the helpers directly from jackin_core::env_model.

Touches (5 files modified, 2 deleted, 1 created):
  - created crates/jackin-core/src/env_model/tests.rs: receives the
    13 tests (the 2 inline  tests from
    jackin-core/src/env_model.rs and the 11 from the deleted
    jackin/src/env_model/tests.rs, deduplicated by function name; the
    dedup is a no-op since the two test sources share no test names)
  - crates/jackin-core/src/env_model.rs: replaces the inline
     block with
    (sibling tests file is the canonical layout; this entry fixes the
    test-layout allowlist by removing the file from the violators list)
  - crates/jackin/src/env_model.rs: deleted (5 lines; the shim was
     + ; the
    binary's app/ layer reads through  directly)
  - crates/jackin/src/env_model/tests.rs: deleted (131 lines; all tests
    moved to the canonical home in jackin-core/src/env_model/tests.rs)
  - crates/jackin/src/lib.rs: removed  line
  - crates/jackin/src/app/config_cmd.rs line 275:
    ->
  - crates/jackin/src/app/workspace_cmd.rs line 555: same repoint
  - test-layout-allowlist.toml: removed the
    entry (no longer a violator)

Decisions:
  - The canonical home is jackin-core (the L0 vocabulary crate). The
    spec's two options were 'jackin-env re-export' or 'jackin-core is
    canonical'; chose the second because env_resolution is not a
    property of the env resolver — it is a vocabulary layer (constants,
    helpers) that the env resolver consumes.
  - The binary's call sites repoint from
    to . The shim deletion is
    what the spec calls 'loser becomes a re-export or is deleted'; the
    shim was a no-op re-export so deletion is the correct choice.
  - The two test files are merged rather than picking one; the
    deduplication is safe because the 2 inline tests from
    jackin-core/src/env_model.rs and the 11 from jackin/src/env_model/
    tests.rs share no test-function name.

Verified:
  - cargo fmt --check clean
  - cargo clippy --workspace --all-targets --all-features --locked -- -D warnings clean
  - cargo nextest run -p jackin-core: passes
  - cargo nextest run -p jackin: 787/787 pass
  - cargo xtask lint: file-size + test-layout + dep-direction all OK
  - grep -r 'crate::env_model' crates/jackin/src/ returns no output
  - behavioral specs runtime-launch and op-picker pass UNMODIFIED

Refs: plans/codebase-health-playbook.md#d2--dedup-env-resolution-homes-p5
Refs: docs/content/docs/roadmap/codebase-health-enforcement.mdx#workstream-w1
