+++ crates/jackin-core/src/env_model/tests.rs
@@ -1,5 +1,19 @@
-//! Tests for `env_model`.
 use super::*;
+
+#[test]
+fn open_links_allowed_accepts_unset_and_non_deny_values() {
+    assert!(open_links_allowed(None));
+    assert!(open_links_allowed(Some("")));
+    assert!(open_links_allowed(Some("allow")));
+    assert!(open_links_allowed(Some("yes")));
+}
+
+#[test]
+fn open_links_allowed_rejects_deny_values() {
+    for value in ["deny", "off", "no"] {
+        assert!(!open_links_allowed(Some(value)));
+    }
+}
 
 #[test]
 fn reserved_runtime_env_vars_covers_every_previously_reserved_name() {
