--- crates/jackin/src/env_model.rs
@@ -1,8 +1,0 @@
-//! Env policy model — re-exported from `jackin-core`.
-//!
-//! The canonical implementation lives in `jackin_core::env_model`.
-
-pub(crate) use jackin_core::env_model::*;
-
-#[cfg(test)]
-mod tests;
