+++ crates/jackin/src/lib.rs
@@ -29,7 +29,6 @@ pub mod derived_image;
 pub(crate) mod diagnostics;
 pub mod docker;
 pub mod docker_client;
-pub(crate) mod env_model;
 pub mod env_resolver;
 pub mod error;
 pub mod instance;
