+++ docs/content/docs/roadmap/codebase-health-enforcement.mdx
@@ -243,7 +243,7 @@ Used by every C/E slice. Pure move; contents are not rewritten.
 ### Phase D — collapse duplicated homes (P5)
 
 - [ ] **D1 — absorb `runtime/image.rs` into `jackin-image`** (also the W5 decomposition of that 2811L file): image build orchestration lives in `jackin-image`; `jackin-runtime` calls it. Splits the file along the way (Dockerfile gen / build / cache-inspection).
-- [ ] **D2 — env-resolution dedup:** one home for the env-resolution types (`jackin-env` vs `jackin-core/env_model`); loser becomes a re-export or is deleted.
+- [x] **D2 — env-resolution dedup:** one home for the env-resolution types (`jackin-env` vs `jackin-core/env_model`); loser becomes a re-export or is deleted.
 - [ ] **D3 — `op_cache` dedup:** one `op_cache` (currently in both `jackin-core` and `jackin-console`).
 - [x] **D4 — resource-naming dedup:** one home for the `*-dind`/`*-net`/volume naming (currently `instance/naming.rs` and `runtime/naming.rs`).
